+++ src/allpress/settings.py
@@ -1,0 +1,10 @@
+
+
+
+POSTGRESQL_SERVER_HOST     = '127.0.0.1'
+POSTGRESQL_DATABASE_NAME   = 'hassan2022cbtest'
+POSTGRESQL_USERNAME        = 'hassan2022cbtest'
+POSTGRESQL_PASSWORD        = 'test0000'
+
+URL_REGEX = '((http|https):\/\/)?((www|ww\d|www\d)\.)?(?=.{5,255})([\w-]{2,63}\.)+\w{2,63}(\/[\w\-._~:?#@!$&\'\(\)*+,;%=]+)*'
+HREF_PARSE_REGEX = '(?<=<a\shref=([\'"]))([\w\-._~:?#@!$&/\'\(\)*+,;%=]+)\1'